# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CarouselSlide {
  /** Slide index across all slides (1-based) */
  slideIndex: number;
  /** Position within its carousel post (1 = cover slide, 2-5 = detail slides) */
  groupPosition: number;
  /** Which carousel post this slide belongs to (1-based) */
  groupIndex: number;
  /** Caption text for this slide */
  text: string;
  /** URL path to the paired photo */
  imageUrl: string;
  /** Original filename of the photo */
  imageName: string;
  /** When true, this slide is a pre-made cover image — render as-is with no text or overlay */
  isCoverImageSlide?: boolean;
}

export interface CarouselResult {
  /** Generated slides — grouped into carousel posts */
  slides: CarouselSlide[];
  /** Total number of slides generated */
  totalSlides: number;
  /** Number of slides per carousel post */
  slidesPerCarousel: number;
  /** Number of carousel posts */
  totalCarousels: number;
  /** Session ID for referencing uploaded images */
  sessionId: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AiSourcePhoto {
  id: number;
  clientName: string;
  uploader: string;
  photoUrl: string;
  notes: string;
  uploadedAt: string;
}

export type AiGeneratedPortraitStatus =
  (typeof AiGeneratedPortraitStatus)[keyof typeof AiGeneratedPortraitStatus];

export const AiGeneratedPortraitStatus = {
  generating: "generating",
  success: "success",
  failed: "failed",
} as const;

export interface AiGeneratedPortrait {
  id: number;
  clientName: string;
  sourcePhotoId: number;
  scenarioId: string;
  prompt: string;
  scrubColor?: string;
  outfitStyle?: string;
  aspectRatio: string;
  originalImageUrl?: string;
  outputImageUrl?: string;
  hasWatermark: boolean;
  status: AiGeneratedPortraitStatus;
  failureReason?: string;
  generatedAt: string;
  savedToLibrary: boolean;
}

export type AiScenarioCategory =
  (typeof AiScenarioCategory)[keyof typeof AiScenarioCategory];

export const AiScenarioCategory = {
  clinical: "clinical",
  lifestyle: "lifestyle",
  brand: "brand",
} as const;

export interface AiScenario {
  id: string;
  name: string;
  category: AiScenarioCategory;
  hasScrubColor: boolean;
  hasOutfitStyle: boolean;
}

export interface AiScenarioConfig {
  id: string;
  scrubColor?: string;
  outfitStyle?: string;
  aspectRatio: string;
}

export interface AiGenerateRequest {
  sourcePhotoId: number;
  clientName?: string;
  /** @maxItems 6 */
  scenarios: AiScenarioConfig[];
}

export type AiCardStateStatus =
  (typeof AiCardStateStatus)[keyof typeof AiCardStateStatus];

export const AiCardStateStatus = {
  idle: "idle",
  generating: "generating",
  success: "success",
  failed: "failed",
  "rate-limited": "rate-limited",
} as const;

export interface AiCardState {
  scenarioId: string;
  status: AiCardStateStatus;
  portraitId?: number;
  originalImageUrl?: string;
  outputImageUrl?: string;
  failureReason?: string;
  retryAfter?: number;
}

export interface AiJobStatus {
  cards: AiCardState[];
}

export type GenerateCarouselBody = {
  /** Photo files to cycle through across slides */
  photos: Blob[];
  /** CSV file — one row per slide, first column is the caption text */
  csv: Blob;
};

export type UploadAiSourcePhotoBody = {
  photo: Blob;
  clientName?: string;
  notes?: string;
};

export type IngestAiSourcePhotoFromUrlBody = {
  photoUrl: string;
  clientName?: string;
  notes?: string;
};

export type GenerateAiPortraits202 = {
  jobId: string;
};

export type SaveAiPortraitToLibraryBody = {
  applyWatermark?: boolean;
};

export type SaveAiPortraitToLibrary200 = {
  success: boolean;
  batchId: number;
  approvalToken: string;
};

export type RegenerateAiPortrait202 = {
  jobId: string;
  portraitId: number;
};