# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: carousel
    description: Carousel post generation
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /carousel/generate:
    post:
      operationId: generateCarousel
      tags: [carousel]
      summary: Generate carousel posts
      description: Upload photos and a CSV where each row is one post (photo + caption). Returns up to 30 posts.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photos:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Photo files to cycle through across slides
                csv:
                  type: string
                  format: binary
                  description: CSV file — one row per slide, first column is the caption text
              required:
                - photos
                - csv
      responses:
        "200":
          description: Generated slide list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarouselResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /carousel/image/{sessionId}/{filename}:
    get:
      operationId: getCarouselImage
      tags: [carousel]
      summary: Get uploaded image
      description: Returns a previously uploaded image by session and filename
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Image file
          content:
            image/*:
              schema:
                type: string
                format: binary
        "404":
          description: Image not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /ai-portrait/source:
    post:
      operationId: uploadAiSourcePhoto
      tags: [ai-portrait]
      summary: Upload a reference source photo
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photo:
                  type: string
                  format: binary
                clientName:
                  type: string
                notes:
                  type: string
              required: [photo]
      responses:
        "200":
          description: Source photo record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSourcePhoto"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /ai-portrait/source-from-url:
    post:
      operationId: ingestAiSourcePhotoFromUrl
      tags: [ai-portrait]
      summary: Ingest a reference source photo from a URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                photoUrl:
                  type: string
                clientName:
                  type: string
                notes:
                  type: string
              required: [photoUrl]
      responses:
        "200":
          description: Source photo record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSourcePhoto"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /ai-portrait/scenarios:
    get:
      operationId: getAiPortraitScenarios
      tags: [ai-portrait]
      summary: Get all available portrait scenarios
      responses:
        "200":
          description: List of scenarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiScenario"

  /ai-portrait/generate:
    post:
      operationId: generateAiPortraits
      tags: [ai-portrait]
      summary: Start a portrait generation job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiGenerateRequest"
      responses:
        "202":
          description: Job accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                required: [jobId]
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /ai-portrait/jobs/{jobId}/status:
    get:
      operationId: getAiPortraitJobStatus
      tags: [ai-portrait]
      summary: Poll the status of a portrait generation job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiJobStatus"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /ai-portrait/{portraitId}/save-to-library:
    post:
      operationId: saveAiPortraitToLibrary
      tags: [ai-portrait]
      summary: Save a generated portrait to the content library
      parameters:
        - name: portraitId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                applyWatermark:
                  type: boolean
      responses:
        "200":
          description: Saved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  batchId:
                    type: integer
                  approvalToken:
                    type: string
                required: [success, batchId, approvalToken]
        "404":
          description: Portrait not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /ai-portrait/{portraitId}/regenerate:
    post:
      operationId: regenerateAiPortrait
      tags: [ai-portrait]
      summary: Retry generation for a failed portrait card
      parameters:
        - name: portraitId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "202":
          description: Regeneration started
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                  portraitId:
                    type: integer
                required: [jobId, portraitId]
        "404":
          description: Portrait not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    CarouselSlide:
      type: object
      properties:
        slideIndex:
          type: integer
          description: Slide index across all slides (1-based)
        groupPosition:
          type: integer
          description: Position within its carousel post (1 = cover slide, 2-5 = detail slides)
        groupIndex:
          type: integer
          description: Which carousel post this slide belongs to (1-based)
        text:
          type: string
          description: Caption text for this slide
        imageUrl:
          type: string
          description: URL path to the paired photo
        imageName:
          type: string
          description: Original filename of the photo
        isCoverImageSlide:
          type: boolean
          description: When true, this slide is a pre-made cover image — render as-is with no text or overlay
      required:
        - slideIndex
        - groupPosition
        - groupIndex
        - text
        - imageUrl
        - imageName

    CarouselResult:
      type: object
      properties:
        slides:
          type: array
          items:
            $ref: "#/components/schemas/CarouselSlide"
          description: Generated slides — grouped into carousel posts
        totalSlides:
          type: integer
          description: Total number of slides generated
        slidesPerCarousel:
          type: integer
          description: Number of slides per carousel post
        totalCarousels:
          type: integer
          description: Number of carousel posts
        sessionId:
          type: string
          description: Session ID for referencing uploaded images
      required:
        - slides
        - totalSlides
        - slidesPerCarousel
        - totalCarousels
        - sessionId

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    AiSourcePhoto:
      type: object
      properties:
        id:
          type: integer
        clientName:
          type: string
        uploader:
          type: string
        photoUrl:
          type: string
        notes:
          type: string
        uploadedAt:
          type: string
          format: date-time
      required: [id, clientName, uploader, photoUrl, notes, uploadedAt]

    AiGeneratedPortrait:
      type: object
      properties:
        id:
          type: integer
        clientName:
          type: string
        sourcePhotoId:
          type: integer
        scenarioId:
          type: string
        prompt:
          type: string
        scrubColor:
          type: string
        outfitStyle:
          type: string
        aspectRatio:
          type: string
        originalImageUrl:
          type: string
        outputImageUrl:
          type: string
        hasWatermark:
          type: boolean
        status:
          type: string
          enum: [generating, success, failed]
        failureReason:
          type: string
        generatedAt:
          type: string
          format: date-time
        savedToLibrary:
          type: boolean
      required: [id, clientName, sourcePhotoId, scenarioId, prompt, aspectRatio, hasWatermark, status, generatedAt, savedToLibrary]

    AiScenario:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
          enum: [clinical, lifestyle, brand]
        hasScrubColor:
          type: boolean
        hasOutfitStyle:
          type: boolean
      required: [id, name, category, hasScrubColor, hasOutfitStyle]

    AiScenarioConfig:
      type: object
      properties:
        id:
          type: string
        scrubColor:
          type: string
        outfitStyle:
          type: string
        aspectRatio:
          type: string
      required: [id, aspectRatio]

    AiGenerateRequest:
      type: object
      properties:
        sourcePhotoId:
          type: integer
        clientName:
          type: string
        scenarios:
          type: array
          items:
            $ref: "#/components/schemas/AiScenarioConfig"
          maxItems: 6
      required: [sourcePhotoId, scenarios]

    AiCardState:
      type: object
      properties:
        scenarioId:
          type: string
        status:
          type: string
          enum: [idle, generating, success, failed, rate-limited]
        portraitId:
          type: integer
        originalImageUrl:
          type: string
        outputImageUrl:
          type: string
        failureReason:
          type: string
        retryAfter:
          type: integer
      required: [scenarioId, status]

    AiJobStatus:
      type: object
      properties:
        cards:
          type: array
          items:
            $ref: "#/components/schemas/AiCardState"
      required: [cards]