# lib/api-zod/src/generated/types/aiCardState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiCardStateStatus } from "./aiCardStateStatus";

export interface AiCardState {
  scenarioId: string;
  status: AiCardStateStatus;
  portraitId?: number;
  originalImageUrl?: string;
  outputImageUrl?: string;
  failureReason?: string;
  retryAfter?: number;
}