# lib/api-zod/src/generated/types/aiGeneratedPortrait.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiGeneratedPortraitStatus } from "./aiGeneratedPortraitStatus";

export interface AiGeneratedPortrait {
  id: number;
  clientName: string;
  sourcePhotoId: number;
  scenarioId: string;
  prompt: string;
  scrubColor?: string;
  outfitStyle?: string;
  aspectRatio: string;
  originalImageUrl?: string;
  outputImageUrl?: string;
  hasWatermark: boolean;
  status: AiGeneratedPortraitStatus;
  failureReason?: string;
  generatedAt: Date;
  savedToLibrary: boolean;
}